backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
openai>=1.6.1
supabase==1.0.4
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
numpy==1.24.3
tiktoken==0.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
